[ALPS07826833] Camsys: Modify buffer handle data length
[Detial]change data type from 32 bits to 64 bits

MTK-Commit-Id: 66ecb005e665be6ec366e6db9ae5bb551844e859

CR-Id: ALPS07826833
Feature: [Android Default] Camera Application Basic Functions

+++ drivers/media/platform/mtk-isp/camsys/isp7sp/cam/mtk_cam-hsf-def.h
@@ -35,7 +35,7 @@ struct mtk_cam_hsf_info {
 	u32 enable_raw;
 	u32 cam_module;
 	u64 chunk_iova;
-	uint32_t chunk_hsfhandle;
+	u64 chunk_hsfhandle;
 };
 
 struct cq_info {
